requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3
aiohttp==3.9.5
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
apscheduler==3.10.4
python-dotenv==1.0.1
loguru==0.7.2
fastapi==0.111.0
uvicorn==0.30.1
